+++ zarrs/src/array/codec/bytes_to_bytes/gdeflate.rs
@@ -28,7 +28,7 @@ pub use gdeflate_codec::GDeflateCodec;
 
 use crate::{
     array::{
-        codec::{Codec, CodecError, CodecPlugin},
+        codec::{Codec, CodecError, CodecPlugin, InvalidBytesLengthError},
         RawBytes,
     },
     metadata::v3::{array::codec::gdeflate, MetadataV3},
@@ -61,10 +61,11 @@ const GDEFLATE_STATIC_HEADER_LENGTH: usize = 2 * size_of::<u64>();
 
 fn gdeflate_decode(encoded_value: &RawBytes<'_>) -> Result<Vec<u8>, CodecError> {
     if encoded_value.len() < GDEFLATE_STATIC_HEADER_LENGTH {
-        return Err(CodecError::UnexpectedChunkDecodedSize(
+        return Err(InvalidBytesLengthError::new(
             encoded_value.len(),
-            GDEFLATE_STATIC_HEADER_LENGTH as u64,
-        ));
+            GDEFLATE_STATIC_HEADER_LENGTH,
+        )
+        .into());
     }
 
     // Decode the static header
@@ -77,10 +78,11 @@ fn gdeflate_decode(encoded_value: &RawBytes<'_>) -> Result<Vec<u8>, CodecError> 
     // Check length of dynamic header
     let dynamic_header_length = num_pages * size_of::<u64>();
     if encoded_value.len() < GDEFLATE_STATIC_HEADER_LENGTH + dynamic_header_length {
-        return Err(CodecError::UnexpectedChunkDecodedSize(
+        return Err(InvalidBytesLengthError::new(
             encoded_value.len(),
-            (GDEFLATE_STATIC_HEADER_LENGTH + dynamic_header_length) as u64,
-        ));
+            GDEFLATE_STATIC_HEADER_LENGTH + dynamic_header_length,
+        )
+        .into());
     }
 
     // Decode the pages