+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Add `array:codec::{InvalidBytesLengthError,InvalidArrayShapeError,InvalidNumberOfElementsError,SubsetOutOfBoundsError}`
 - Add `ArraySubset::inbounds_shape()` (matches the old `ArraySubset::inbounds` behaviour)
 
 ### Changed
+- **Breaking**: `CodecError` enum changes:
+  - Change `CodecError::UnexpectedChunkDecodedSize` to an `InvalidBytesLengthError`
+  - Add `CodecError::{InvalidArrayShape,InvalidNumberOfElements,SubsetOutOfBounds}`
 - **Breaking**: change `ArraySubset::inbounds` to take another subset rather than a shape
 
 ## [0.19.1] - 2025-01-19